+++ app/Http/Controllers/KarmaController.php
@@ -6,6 +6,7 @@ use App\Food;
 use App\Karma;
 use App\KarmaFood;
 use App\KarmaSpecificValue;
+use App\Solution;
 use Illuminate\Http\Request;
 
 class KarmaController extends Controller
@@ -74,7 +75,14 @@ class KarmaController extends Controller
                                 }
                             }
                             else{
-                                $KarmaSpecific[$j] =$KarmaSpecific[$j]+ $specificsValues[$k] * $values[$i];
+                                if ($k ==14)
+                                {
+                                    $KarmaSpecific[$j] =$KarmaSpecific[$j]+( $specificsValues[$k] * $values[$i] /1000);
+                                }
+                                else{
+                                    $KarmaSpecific[$j] =$KarmaSpecific[$j]+ $specificsValues[$k] * $values[$i];
+                                }
+
                             }
                             $k =$k+1;
                         }
@@ -92,20 +100,20 @@ class KarmaController extends Controller
         }
         $karmaSpecificValue = new KarmaSpecificValue();
         $karmaSpecificValue->karma_id = $karma->id;
-        $karmaSpecificValue->Km =10 * $KarmaSpecific[0];
-        $karmaSpecificValue->Hp =10 * $KarmaSpecific[1];
-        $karmaSpecificValue->Enerji =1000 * $KarmaSpecific[2];
-        $karmaSpecificValue->Lif =10 * $KarmaSpecific[3];
-        $karmaSpecificValue->Kul = 10 * $KarmaSpecific[4];
-        $karmaSpecificValue->Karbonhidrat =10 * $KarmaSpecific[5];
-        $karmaSpecificValue->Kalsiyum =10 *$KarmaSpecific[6];
-        $karmaSpecificValue->Fosfor =10 * $KarmaSpecific[7];
+        $karmaSpecificValue->Km = $KarmaSpecific[0] /100 ;
+        $karmaSpecificValue->Hp =$KarmaSpecific[1] /100 ;
+        $karmaSpecificValue->Enerji = $KarmaSpecific[2];
+        $karmaSpecificValue->Lif = $KarmaSpecific[3] /100 ;
+        $karmaSpecificValue->Kul =  $KarmaSpecific[4] /100 ;
+        $karmaSpecificValue->Karbonhidrat = $KarmaSpecific[5] /100 ;
+        $karmaSpecificValue->Kalsiyum = $KarmaSpecific[6] /100 ;
+        $karmaSpecificValue->Fosfor = $KarmaSpecific[7] /100 ;
         $karmaSpecificValue->Ca_p = $KarmaSpecific[8];
-        $karmaSpecificValue->Meganizyum =10000* $KarmaSpecific[9];
-        $karmaSpecificValue->Sodyum =10000* $KarmaSpecific[10];
-        $karmaSpecificValue->Taurin =10* $KarmaSpecific[11];
-        $karmaSpecificValue->Yag = 10*$KarmaSpecific[12];
-        $karmaSpecificValue->LinoliekAsit =10 * $KarmaSpecific[13];
+        $karmaSpecificValue->Meganizyum =10* $KarmaSpecific[9];
+        $karmaSpecificValue->Sodyum =10* $KarmaSpecific[10];
+        $karmaSpecificValue->Taurin = $KarmaSpecific[11] /100 ;
+        $karmaSpecificValue->Yag = $KarmaSpecific[12] /100 ;
+        $karmaSpecificValue->LinoliekAsit = $KarmaSpecific[13] /100 ;
         $karmaSpecificValue->Fiyat = $KarmaSpecific[14];
         $karmaSpecificValue->save();
         toast(__('Mixture Added Successfully'),'success');
@@ -195,7 +203,14 @@ class KarmaController extends Controller
                                 }
                             }
                             else{
-                                $KarmaSpecific[$j] =$KarmaSpecific[$j]+ $specificsValues[$k] * $values[$i];
+                                if ($k ==14)
+                                {
+                                    $KarmaSpecific[$j] =$KarmaSpecific[$j]+( $specificsValues[$k] * $values[$i] /1000);
+                                }
+                                else{
+                                    $KarmaSpecific[$j] =$KarmaSpecific[$j]+( $specificsValues[$k] * $values[$i]);
+                                }
+
                             }
                             $k =$k+1;
                         }
@@ -213,20 +228,20 @@ class KarmaController extends Controller
         }
         $karmaSpecificValue = new KarmaSpecificValue();
         $karmaSpecificValue->karma_id = $karma->id;
-        $karmaSpecificValue->Km =10 * $KarmaSpecific[0];
-        $karmaSpecificValue->Hp =10 * $KarmaSpecific[1];
-        $karmaSpecificValue->Enerji =1000 * $KarmaSpecific[2];
-        $karmaSpecificValue->Lif =10 * $KarmaSpecific[3];
-        $karmaSpecificValue->Kul = 10 * $KarmaSpecific[4];
-        $karmaSpecificValue->Karbonhidrat =10 * $KarmaSpecific[5];
-        $karmaSpecificValue->Kalsiyum =10 *$KarmaSpecific[6];
-        $karmaSpecificValue->Fosfor =10 * $KarmaSpecific[7];
+        $karmaSpecificValue->Km = $KarmaSpecific[0] /100 ;
+        $karmaSpecificValue->Hp =$KarmaSpecific[1] /100 ;
+        $karmaSpecificValue->Enerji = $KarmaSpecific[2];
+        $karmaSpecificValue->Lif = $KarmaSpecific[3] /100 ;
+        $karmaSpecificValue->Kul =  $KarmaSpecific[4] /100 ;
+        $karmaSpecificValue->Karbonhidrat = $KarmaSpecific[5] /100 ;
+        $karmaSpecificValue->Kalsiyum = $KarmaSpecific[6] /100 ;
+        $karmaSpecificValue->Fosfor = $KarmaSpecific[7] /100 ;
         $karmaSpecificValue->Ca_p = $KarmaSpecific[8];
-        $karmaSpecificValue->Meganizyum =10000* $KarmaSpecific[9];
-        $karmaSpecificValue->Sodyum =10000* $KarmaSpecific[10];
-        $karmaSpecificValue->Taurin =10* $KarmaSpecific[11];
-        $karmaSpecificValue->Yag = 10*$KarmaSpecific[12];
-        $karmaSpecificValue->LinoliekAsit =10 * $KarmaSpecific[13];
+        $karmaSpecificValue->Meganizyum =10* $KarmaSpecific[9];
+        $karmaSpecificValue->Sodyum =10* $KarmaSpecific[10];
+        $karmaSpecificValue->Taurin = $KarmaSpecific[11] /100 ;
+        $karmaSpecificValue->Yag = $KarmaSpecific[12] /100 ;
+        $karmaSpecificValue->LinoliekAsit = $KarmaSpecific[13] /100 ;
         $karmaSpecificValue->Fiyat = $KarmaSpecific[14];
         $karmaSpecificValue->save();
         toast(__('Mixture Updated Successfully'),'success');
@@ -242,15 +257,30 @@ class KarmaController extends Controller
     public function destroy($id)
     {
         $karma = Karma::findOrFail($id);
-        $karma->delete();
-        $KarmaFoods = KarmaFood::where('karma_id',$id)->get();
-        $KarmaSpecific = KarmaSpecificValue::where('karma_id',$id);
-        $KarmaSpecific->delete();
-        foreach ($KarmaFoods as $KarmaFood)
+        $solution = Solution::where('karma_id',$id)->get()->toArray();
+        if($solution != null)
         {
-            $KarmaFood->delete();
+            toast(__("Please Delete the Related Solution That has This Mixture Firstly ..."),'error');
+            return redirect()->route('solution.index');
         }
-        toast(__('Mixture Deleted Successfully'),'info');
-        return redirect()->route('karma.index');
+        if(auth()->user()->id == 1 || auth()->user()->id == $karma->user_id)
+        {
+            $karma->delete();
+            $KarmaFoods = KarmaFood::where('karma_id',$id)->get();
+            $KarmaSpecific = KarmaSpecificValue::where('karma_id',$id);
+            $KarmaSpecific->delete();
+            foreach ($KarmaFoods as $KarmaFood)
+            {
+                $KarmaFood->delete();
+            }
+            toast(__('Mixture Deleted Successfully'),'info');
+            return redirect()->route('karma.index');
+        }
+        else
+        {
+            toast(__("You can't Delete this Mixture ..."),'error');
+            return redirect()->route('animal.index');
+        }
+
     }
 }
